Add GET endpoint option
One of the powers of persisted queries is that you can leverage CDN's to offload caching.  It's a opt in obviously, since it's for user's who understand the dangers and complexities of introducing such a optimization (settings appropriate cache-control headers, ensuring query being used only contains data that can *fully* be cached, ect).